Pulled the `EmailSender` class into its own file & renamed /email-sender endpoint to /contact-support
